requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
folium
streamlit-folium
mlflow
umap-learn
plotly
scipy
